requirements.txt
psycopg2-binary>=2.9.6
requests>=2.28.0
beautifulsoup4>=4.11.0
pyyaml>=6.0 